requirements.txt
aiogram==3.13.1
yt-dlp
python-dotenv
aiofiles